+++ Server/wiselaundry/src/main/java/com/ssafy/wiselaundry/domain/user/db/repository/UserRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
-    User findByUserId(int userId);
+public interface UserRepository extends JpaRepository<User, Long> {
+    User findByUserId(Long userId);
     User findByUserEmail(String userEmail);
     User findByUserNick(String userNick);
 }
